CV python automation script
CV python automation script to process CV csv data and plot the result. It will also ask for prompts to calculate the HOMO and LUMO levels from inputted onsets as intersections of line inputs. It will output a picture of the graph, calculations, and an excel file with values if user wants to plot in excel as well.